3801 - Replaced the deprecated StatisticBoxes for feral druid
