+++ src/parser/druid/feral/modules/core/HitCountAoE.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import SpellIcon from 'common/SpellIcon';
 import Analyzer, { SELECTED_PLAYER } from 'parser/core/Analyzer';
-import StatisticBox from 'interface/others/StatisticBox';
+import BoringSpellValueText from 'interface/statistics/components/BoringSpellValueText';
+import Statistic from 'interface/statistics/Statistic';
 import Events from 'parser/core/Events';
 
 // time after a cast in which direct damage from the spellId will be associated with the cast
@@ -23,7 +23,7 @@ class HitCountAoE extends Analyzer {
   lastCastEvent = null;
   lastCastHits = 0;
 
-  constructor(options){
+  constructor(options) {
     super(options);
     this.addEventListener(Events.cast.by(SELECTED_PLAYER).spell(this.constructor.spell), this.onCast);
     this.addEventListener(Events.damage.by(SELECTED_PLAYER).spell(this.constructor.spell), this.onDamage);
@@ -39,7 +39,7 @@ class HitCountAoE extends Analyzer {
 
   onDamage(event) {
     if (event.tick ||
-        !this.lastCastEvent || ((event.timestamp - this.lastCastEvent.timestamp) > DAMAGE_WINDOW)) {
+      !this.lastCastEvent || ((event.timestamp - this.lastCastEvent.timestamp) > DAMAGE_WINDOW)) {
       // only interested in direct damage from the spellId shortly after cast
       return;
     }
@@ -91,10 +91,8 @@ class HitCountAoE extends Analyzer {
       return null;
     }
     return (
-      <StatisticBox
-        icon={<SpellIcon id={this.constructor.spell.id} />}
-        value={`${this.averageTargetsHit.toFixed(1)}`}
-        label={`Average targets hit by ${this.constructor.spell.name}`}
+      <Statistic
+        size="flexible"
         tooltip={(
           <>
             You used {this.constructor.spell.name} <strong>{this.casts}</strong> time{(this.casts === 1) ? '' : 's'}.<br />
@@ -105,7 +103,13 @@ class HitCountAoE extends Analyzer {
           </>
         )}
         position={statisticPosition}
-      />
+      >
+        <BoringSpellValueText spell={this.constructor.spell}>
+          <>
+            {this.averageTargetsHit.toFixed(1)} <small>average targets hit</small>
+          </>
+        </BoringSpellValueText>
+      </Statistic>
     );
   }
 }
