+++ src/parser/druid/feral/modules/talents/Predator.js
@@ -1,9 +1,10 @@
 import React from 'react';
 import Analyzer, { SELECTED_PLAYER } from 'parser/core/Analyzer';
 import SPELLS from 'common/SPELLS';
-import SpellIcon from 'common/SpellIcon';
 import SpellLink from 'common/SpellLink';
-import StatisticBox, { STATISTIC_ORDER } from 'interface/others/StatisticBox';
+import STATISTIC_ORDER from 'interface/others/STATISTIC_ORDER';
+import BoringSpellValueText from 'interface/statistics/components/BoringSpellValueText';
+import Statistic from 'interface/statistics/Statistic';
 
 import { i18n } from '@lingui/core';
 import { t } from '@lingui/macro';
@@ -43,6 +44,7 @@ class Predator extends Analyzer {
   get extraCasts() {
     return Math.max(0, this.totalCasts - this.baseCasts);
   }
+
   get extraCastsPerMinute() {
     return (this.extraCasts / this.owner.fightDuration) * 1000 * 60;
   }
@@ -76,13 +78,17 @@ class Predator extends Analyzer {
     const hadExtraCasts = <>Your Predator talent allowed you to use Tiger's Fury at least <strong>{this.extraCasts}</strong> extra time{this.extraCasts !== 1 ? 's' : ''}. Without it you would have had time for <strong>{this.baseCasts}</strong> cast{this.baseCasts !== 1 ? 's' : ''} but with it you were able to use Tiger's Fury <strong>{this.totalCasts}</strong> time{this.totalCasts !== 1 ? 's' : ''}.{this.earlyCasts > 0 && earlyCastsComment}</>;
     const noExtraCasts = <>Your Predator talent didn't allow you to cast more Tiger's Fury overall than you would have been able to without it, with the fight lasting long enough for all <strong>{this.totalCasts}</strong> of your cast{this.totalCasts !== 1 ? 's' : ''}. Either there were no enemies dying with your bleeds on them during this fight or you didn't make use of Tiger's Fury when it came off cooldown.{this.earlyCasts > 0 && earlyCastsComment}</>;
     return (
-      <StatisticBox
-        icon={<SpellIcon id={SPELLS.PREDATOR_TALENT.id} />}
-        value={this.extraCastsPerMinute.toFixed(2)}
-        label="Extra Tiger's Fury casts per minute"
+      <Statistic
+        size="flexible"
         tooltip={this.extraCasts > 0 ? hadExtraCasts : noExtraCasts}
         positon={STATISTIC_ORDER.OPTIONAL(3)}
-      />
+      >
+        <BoringSpellValueText spell={SPELLS.PREDATOR_TALENT}>
+          <>
+            {this.extraCastsPerMinute.toFixed(2)} <small>extra casts per minute</small>
+          </>
+        </BoringSpellValueText>
+      </Statistic>
     );
   }
 }
